Move cnn code to utilities file
